% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{build_example}
\alias{build_example}
\title{Generate a plot}
\usage{
build_example()
}
\value{
Returns an example plot which uses the package
}
\description{
`build_example()` Generate an example of what is in the package
}
\examples{
if (interactive()) {
    build_example()
}
}
