% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{dren_scatter}
\alias{dren_scatter}
\title{Generate a scatter plot}
\usage{
dren_scatter(
  title = "My Scatter",
  x_title = "MPG",
  y_title = "WT",
  font = "serif",
  font_size = 16,
  point_size = 1,
  add_lm = TRUE,
  lm_colour = "black",
  lm_size = 2
)
}
\arguments{
\item{title}{The main title for the plot}

\item{x_title}{The main label for X axis}

\item{y_title}{The main label for Y axis}

\item{font}{The preferred font}

\item{font_size}{The font size}

\item{point_size}{The point size}

\item{add_lm}{Add a best fit line}

\item{lm_colour}{Color of regression line}

\item{lm_size}{Size of regression line}
}
\value{
Returns an example plot which uses the package
}
\description{
`dren_scatter()` Generate an example of what is in the package
}
\examples{
if (interactive()) {
  ggplot(mtcars, aes(x = mpg, y = wt)) +
    dren_scatter()
}
}
