% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dren_theme.R
\name{dren_theme}
\alias{dren_theme}
\title{Generate the main theme}
\usage{
dren_theme()
}
\description{
`dren_theme()` Generate the main theme
}
\examples{
if (interactive()) {
    mtcars \%>\%
     ggplot(aes(x = mpg, y = wt)) +
     geom_point() +
     dren_theme()
}
}
