% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{lincoln_weather}
\alias{lincoln_weather}
\title{lincoln_weather}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 366 rows and 24 columns.
}
\usage{
lincoln_weather
}
\description{
lincoln_weather
}
\keyword{datasets}
