% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_moons.R
\name{make_moons}
\alias{make_moons}
\title{Generate a dataset}
\usage{
make_moons(n_samples = 20000L, shuffle = TRUE, noise = 0.55)
}
\value{
Returns a moon datast
}
\description{
Generate a dataset
}
\examples{
if (interactive()) {
    make_moons()
}
}
