% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{moons_dataset}
\alias{moons_dataset}
\title{moons_dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 20000 rows and 3 columns.
}
\usage{
moons_dataset
}
\description{
moons_dataset
}
\keyword{datasets}
