% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{random_moons}
\alias{random_moons}
\title{random_moons}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7377 rows and 5 columns.
}
\usage{
random_moons
}
\description{
random_moons
}
\keyword{datasets}
