% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{statedata}
\alias{statedata}
\title{statedata}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 51 rows and 6 columns.
}
\usage{
statedata
}
\description{
statedata
}
\keyword{datasets}
