% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{unemployment}
\alias{unemployment}
\title{unemployment}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 3217 rows and 9 columns.
}
\usage{
unemployment
}
\description{
unemployment
}
\keyword{datasets}
