% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{dren_scatter}
\alias{dren_scatter}
\title{Generate a scatter plot}
\usage{
dren_scatter(
  title = "My Scatter",
  x_title = "MPG",
  y_title = "WT",
  font = "serif",
  font_size = 16
)
}
\value{
Returns an example plot which uses the package
}
\description{
`dren_scatter()` Generate an example of what is in the package
}
\examples{
if (interactive()) {
  ggplot(mtcars, aes(x = mpg, y = wt)) +
    dren_scatter()
}
}
